feat(src): use vue2mapbox-gl
