+++ package.json
@@ -30,6 +30,7 @@
     "vue-echarts": "^6.2.3",
     "vue-mapbox": "^0.4.1",
     "vue-router": "^3.1.3",
+    "vue2mapbox-gl": "^0.13.1",
     "vuetify": "^2.3.13",
     "vuex": "^3.1.1"
   },
