+++ src/main.js
@@ -1,6 +1,7 @@
 import Vue from 'vue';
 import App from '@/App.vue';
 import vuetify from '@/plugins/vuetify';
+import vue2mapbox from '@/plugins/vue2mapbox';
 import 'mapbox-gl/dist/mapbox-gl.css';
 import store from '@/store';
 import router from '@/router';
@@ -11,6 +12,7 @@ Vue.config.productionTip = false;
 
 new Vue({
   vuetify,
+  vue2mapbox,
   store,
   router,
   render: h => h(App),
