+++ src/plugins/vue2mapbox.js
@@ -1,0 +1,9 @@
+import Vue from 'vue';
+import Vue2MapboxGL from 'vue2mapbox-gl';
+
+import 'mapbox-gl/dist/mapbox-gl.css';
+import '@mapbox/mapbox-gl-geocoder/dist/mapbox-gl-geocoder.css';
+
+Vue.use(Vue2MapboxGL);
+
+export default Vue2MapboxGL;
